refactor(supervisor): PR-4 commit 1 — supervisor.py 1431 → supervisor/ 6 文件子包
按职责切叶子 helpers 到独立模块，Supervisor 主类保单类不拆（行为零变更）：

  slot.py         _Slot dataclass + SLOT_TRAIN/DATA 常量
  cmd_builder.py  默认 cmd builder + monitor_state_path + EVENT 协议常量 + type aliases
  finalizer.py    _maybe_finalize_version（task 终态 → version.status，ADR-0007 §11.3-B）
  process.py      _kill_process_tree（跨平台杀进程树）
  core.py         Supervisor 主类整搬 1162 行（37 method 全部保留）
  __init__.py     re-export shim

**Supervisor 主类不拆**的决策（PR-2 schema TrainingConfig 单类同款判断）：
37 个 method 全部 read/write 共享 self 字段（_slots / _daemon_* / _stop /
_db_path），状态耦合极高且缺乏清晰子域边界 —— 拆 Mixin/helper class 反而
增加未来扩展（pause/resume PR-3、多 GPU 槽位、quota）跨文件改 self 状态
的成本。叶子 helper 拆出降低顶层 namespace 心智负担即可。

兼容垫片：
- __init__.py 全 public name re-export，旧 `from studio.supervisor import
  Supervisor / _Slot / _default_cmd_builder / _maybe_finalize_version` 透明
- monkeypatch path 兼容：暴露 `_secrets` / `subprocess` 作 supervisor 包
  attribute，tests 用的 `monkeypatch.setattr("studio.supervisor._secrets.load",
  X)` / `"studio.supervisor.subprocess.Popen"` 仍能 hit module 单例。__init__.py
  从 .core re-export 这俩名让 lookup 命中真实 module（Python 模块单例 →
  patch 同步影响 core 内调用）

回归：tests/test_supervisor* + test_inference_daemon + test_resume_path +
test_daemon_gpu_yield = 82/82 全绿；PR-1 安全网 144/144 全绿。
